numcodecs_wasm_host_reproducible

Struct ReproducibleWasmCodec

pub struct ReproducibleWasmCodec<E: WasmEngine>
where
    Store<(), ReproducibleEngine<E>>: Send,
{ /* private fields */ }

Codec instantiated inside a WebAssembly component.

The codec is loaded such that its execution is reproducible across any platform. Importantly, each codec owns its own component instance such that two codecs cannot interfere.

Implementations

impl<E: WasmEngine> ReproducibleWasmCodec<E>
where Store<(), ReproducibleEngine<E>>: Send,

pub fn try_clone(&self) -> Result<Self, Error>

Try cloning the codec by recreating it from its configuration.

ReproducibleWasmCodec implements Clone by calling this method and panicking if it fails.

Errors

Errors if serializing the codec configuration, constructing the new codec, or interacting with the component fails.

pub fn try_drop(self) -> Result<(), ReproducibleWasmCodecError>

Try dropping the codec.

ReproducibleWasmCodec implements Drop by calling this method and ignoring any errors.

Errors

Errors if dropping the codecโs resource or component, or interacting with the component fails.

pub fn instruction_counter( &self, ) -> Result<Wrapping<u64>, ReproducibleWasmCodecError>

Read the codecโs instruction counter, which is based on the number of WebAssembly bytecode instructions executed.

The instruction counter is never reset and wraps around. Comparisons of the counter values before and after, e.g. a call to Self::encode, should thus use wrapping arithmetic.

Errors

Errors if interacting with the component fails.